requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
influxdb-client==1.15.0
psutil==5.8.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
Rx==3.1.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
